backend/requirements.txt
fastapi[all]
uvicorn[standard]
sqlalchemy
psycopg2
alembic
python-jose[cryptography]
passlib[bcrypt]
keras
numpy
tensorflow
pandas
